+++ packages/vue/src/carousel/carousel-context.ts
@@ -1,9 +1,7 @@
-import { type connect } from '@zag-js/carousel'
-import { type ComputedRef, type UnwrapRef } from 'vue'
 import { createContext } from '../context'
 import { type UseCarouselReturn } from './use-carousel'
 
-export const [CarouselProvider, useCarouselContext] =
-  createContext<ComputedRef<ReturnType<typeof connect>>>('CarouselContext')
+export type CarouselContext = UseCarouselReturn
 
-export type CarouselContext = UnwrapRef<UseCarouselReturn>
+export const [CarouselProvider, useCarouselContext] =
+  createContext<UseCarouselReturn>('CarouselContext')