+++ packages/vue/src/carousel/carousel-control.tsx
@@ -1,5 +1,4 @@
-import { mergeProps } from '@zag-js/vue'
-import { computed, defineComponent } from 'vue'
+import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import { carouselAnatomy } from './carousel.anatomy'
 
@@ -8,8 +7,10 @@ export type CarouselControlProps = HTMLArkProps<'div'>
 export const CarouselControl = defineComponent({
   name: 'CarouselControl',
   setup(_, { slots, attrs }) {
-    const mergedProps = computed(() => mergeProps(attrs, carouselAnatomy.build().control.attrs))
-
-    return () => <ark.div {...mergedProps.value}>{slots.default?.()}</ark.div>
+    return () => (
+      <ark.div {...carouselAnatomy.build().control.attrs} {...attrs}>
+        {slots.default?.()}
+      </ark.div>
+    )
   },
 })