+++ packages/vue/src/carousel/carousel-viewport.tsx
@@ -1,11 +1,10 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import type { ComponentWithProps } from '../utils'
 import { useCarouselContext } from './carousel-context'
 
 export type CarouselViewportProps = HTMLArkProps<'div'>
 
-export const CarouselViewport: ComponentWithProps<CarouselViewportProps> = defineComponent({
+export const CarouselViewport = defineComponent({
   name: 'CarouselViewport',
   setup(_, { slots, attrs }) {
     const api = useCarouselContext()