+++ packages/vue/src/carousel/carousel.tsx
@@ -1,47 +1,18 @@
-import { defineComponent, type PropType } from 'vue'
+import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import type { Assign } from '../types'
-import type { ComponentWithProps } from '../utils'
 import { CarouselProvider } from './carousel-context'
-import { useCarousel, type UseCarouselContext } from './use-carousel'
+import { emits, props } from './carousel.props'
+import { useCarousel, type UseCarouselProps } from './use-carousel'
 
-export type CarouselProps = Assign<HTMLArkProps<'div'>, UseCarouselContext>
+export type CarouselProps = Assign<HTMLArkProps<'div'>, UseCarouselProps>
 
-export const Carousel: ComponentWithProps<CarouselProps> = defineComponent({
+export const Carousel = defineComponent({
   name: 'Carousel',
-  props: {
-    align: {
-      type: String as PropType<CarouselProps['align']>,
-    },
-    dir: {
-      type: String as PropType<CarouselProps['dir']>,
-    },
-    getRootNode: {
-      type: Function as PropType<CarouselProps['getRootNode']>,
-    },
-    id: {
-      type: String as PropType<CarouselProps['id']>,
-    },
-    index: {
-      type: Number as PropType<CarouselProps['index']>,
-    },
-    loop: {
-      type: Boolean as PropType<CarouselProps['loop']>,
-    },
-    orientation: {
-      type: String as PropType<CarouselProps['orientation']>,
-    },
-    slidesPerView: {
-      type: Number as PropType<CarouselProps['slidesPerView']>,
-    },
-    spacing: {
-      type: String as PropType<CarouselProps['spacing']>,
-    },
-  },
-  emits: ['slide-change'],
+  props,
+  emits,
   setup(props, { slots, attrs, emit }) {
-    const api = useCarousel(emit, props)
-
+    const api = useCarousel(props, emit)
     CarouselProvider(api)
 
     return () => (