+++ packages/vue/src/carousel/use-carousel.ts
@@ -1,28 +1,27 @@
 import * as carousel from '@zag-js/carousel'
 import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
-import { computed, reactive, type ComputedRef } from 'vue'
+import { computed, ref, type ComputedRef } from 'vue'
+import { useEnvironmentContext } from '../environment'
 import type { Optional } from '../types'
-import { useId } from '../utils'
+import { generateEventMap, useId } from '../utils'
+import { emits } from './carousel.props'
 
-export type UseCarouselContext = Optional<carousel.Context, 'id'>
+export type UseCarouselProps = Optional<carousel.Context, 'id'>
 export type UseCarouselReturn = ComputedRef<carousel.Api<PropTypes>>
 
-export const useCarousel = (
-  emit: CallableFunction,
-  context: UseCarouselContext,
-): UseCarouselReturn => {
-  const reactiveContext = reactive(context)
+export const useCarousel = (props: UseCarouselProps, emit: CallableFunction): UseCarouselReturn => {
+  const getRootNode = useEnvironmentContext()
+  const context = ref(props)
+  const eventMap = generateEventMap(emits, emit)
 
   const [state, send] = useMachine(
     carousel.machine({
-      ...reactiveContext,
-      getRootNode: reactiveContext.getRootNode,
-      id: reactiveContext.id || useId().value,
-      onSlideChange(details) {
-        emit('slide-change', details)
-      },
+      id: useId().value,
+      getRootNode,
+      ...context.value,
+      ...eventMap,
     }),
+    { context },
   )
-
   return computed(() => carousel.connect(state.value, send, normalizeProps))
 }