+++ packages/vue/CHANGELOG.md
@@ -14,6 +14,7 @@ description: All notable changes to this project will be documented in this file
 ### Changed
 
 - Revised `Accordion` component
+- Revised `Carousel` component
 - Revised `Pagination` component
 - Revised `Splitter` component
 - Revised `Tooltip` component